Move the location of the onboarding component
- Move the file to be in the main component folder.
