+++ client/src/components/OnBoarding.tsx
